// src/modules/client-adm/usecase/find-all-client/find-all-client.usecase.dto.ts
export interface FindAllClientUseCaseOutputDto {
  clients: {
    id: string
    name: string
    email: string
    document: string
    address: {
      street: string
      number: string
      complement: string
      city: string
      state: string
      zipCode: string
    }
    createdAt: Date
    updatedAt: Date
  }[]
}